+++ src/frontend/packages/core/src/core/permissions/current-user-permissions.config.ts
@@ -1,3 +1,5 @@
+import { PermissionValues } from '../../../../store/src/selectors/current-user-role.selectors';
+
 export type PermissionConfigType = PermissionConfig[] | PermissionConfig | PermissionConfigLink;
 export interface IPermissionConfigs {
   [permissionString: string]: PermissionConfigType;
@@ -6,7 +8,6 @@ export interface IPermissionConfigs {
 export type PermissionTypes = string;
 export type CurrentUserPermissions = string;
 export type ScopeStrings = string;
-export type PermissionValues = string;
 export class PermissionConfig {
   constructor(
     public type: PermissionTypes,
