+++ src/frontend/packages/store/src/entity-catalog/entity-catalog-entity/entity-catalog-entity.types.ts
@@ -5,13 +5,20 @@ import { PaginationObservables } from '../../reducers/pagination-reducer/paginat
 import { PaginatedAction } from '../../types/pagination.types';
 import { OrchestratedActionBuilders, OrchestratedActionCoreBuilders } from '../action-orchestrator/action-orchestrator';
 
+export type NonOptionalKeys<T extends object> = Exclude<{
+  [K in keyof T]: T extends Record<K, T[K]>
+  ? K
+  : never
+}[keyof T], undefined>
+
+
 /**
  * Remove keys such as typed indexes  (i.e. [key: string])
  * For magic see
  *  - https://github.com/Microsoft/TypeScript/issues/25987#issuecomment-441224690
  *  - https://github.com/Microsoft/TypeScript/issues/12215#issuecomment-414808995
  */
-type KnownKeys<T> = {
+export type KnownKeys<T> = {
   [K in keyof T]: string extends K ? never : number extends K ? never : K
 } extends { [_ in keyof T]: infer U } ? ({} extends U ? never : U) : never;
